+++ src/parser/inline/autolink/url.test.ts
@@ -83,6 +83,7 @@ describe('Unit: parser/inline/autolink/url', () => {
       assert.deepStrictEqual(inspect(parser('http://a:80')), [['<a href="http://a:80" rel="noopener" target="_blank">http://a:80</a>'], '']);
       assert.deepStrictEqual(inspect(parser('http://a.b')), [['<a href="http://a.b" rel="noopener" target="_blank">http://a.b</a>'], '']);
       assert.deepStrictEqual(inspect(parser('http://a?b=c+d&\\#(a(b)()(c)d)')), [['<a href="http://a?b=c+d&amp;\\#(a(b)()(c)d)" rel="noopener" target="_blank">http://a?b=c+d&amp;\\#(a(b)()(c)d)</a>'], '']);
+      assert.deepStrictEqual(inspect(parser(`http://a?#${encodeURIComponent(':/[]()<>?#=& ')}`)), [['<a href="http://a?#%3A%2F%5B%5D()%3C%3E%3F%23%3D%26%20" rel="noopener" target="_blank">http://a?#%3A%2F[]()&lt;&gt;%3F%23%3D%26%20</a>'], '']);
       assert.deepStrictEqual(inspect(parser('http://[::]')), [['<a href="http://[::]" rel="noopener" target="_blank">http://[::]</a>'], '']);
     });