+++ src/parser/inline/link.test.ts
@@ -74,6 +74,7 @@ describe('Unit: parser/inline/link', () => {
       assert.deepStrictEqual(inspect(parser('[a ]()')), [['<a href="" rel="noopener">a </a>'], '']);
       assert.deepStrictEqual(inspect(parser('[ a]()')), [['<a href="" rel="noopener"> a</a>'], '']);
       assert.deepStrictEqual(inspect(parser('[ a ]()')), [['<a href="" rel="noopener"> a </a>'], '']);
+      assert.deepStrictEqual(inspect(parser(`[](?#${encodeURIComponent(':/[]()<>?#=& ')})`)), [['<a href="?#%3A%2F%5B%5D()%3C%3E%3F%23%3D%26%20" rel="noopener">?#%3A%2F[]()&lt;&gt;%3F%23%3D%26%20</a>'], '']);
     });
 
     it('tel', () => {