+++ src/parser/string/url.test.ts
@@ -1,4 +1,4 @@
-﻿import { sanitize } from './url';
+﻿import { sanitize, decode } from './url';
 
 describe('Unit: parser/text/url', () => {
   describe('sanitiza', () => {
@@ -19,6 +19,17 @@ describe('Unit: parser/text/url', () => {
       assert(sanitize('vbscript:alert)') === '');
       assert(sanitize('data:text/html;base64,PHNjcmlwdD5hbGVydCgnWFNTJyk8L3NjcmlwdD4K)') === '');
       assert(sanitize('any:alert)') === '');
+    });
+
+  });
+
+  describe('decode', () => {
+    it('invalid', () => {
+      assert(decode('%FF') === '%FF');
+    });
+
+    it('valid', () => {
+      assert(decode(' %20') === '%20%20');
     });
 
   });